+++ src/main/java/pl/indexpz/iStat/controller/VehicleRecordController/VehicleRecordDeleteController.java
@@ -7,10 +7,11 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import pl.indexpz.iStat.domain.model.Vehicle;
 import pl.indexpz.iStat.domain.model.VehicleRecord;
 import pl.indexpz.iStat.domain.service.VehicleRecordService;
 import pl.indexpz.iStat.domain.service.VehicleService;
+
+import java.util.Optional;
 
 @Controller
 @RequestMapping("/vehicle-records/delete")
@@ -22,18 +23,23 @@ public class VehicleRecordDeleteController {
     private final VehicleService vehicleService;
 
     @GetMapping
-    public String prepareDelete(Long id, Long vehicleId, Model model){
-        VehicleRecord vehicleRecord = vehicleRecordService.getVehicleRecordById(id);
-//        log.info(vehicleRecord.toString());
-        model.addAttribute("vehicleRecord", vehicleRecord);
-        model.addAttribute("vehicle", vehicleService.getVehicleById(vehicleId));
+    public String prepareDelete(Long vehicleRecordId, Long vehicleId, Model model){
+        Optional<VehicleRecord> optionalVehicleRecord = vehicleRecordService.getVehicleRecordById(vehicleRecordId);
+        if(optionalVehicleRecord.isPresent()){
+            VehicleRecord vehicleRecord = optionalVehicleRecord.get();
+            model.addAttribute("vehicleRecord", vehicleRecord);
+            model.addAttribute("vehicle", vehicleService.getVehicleById(vehicleId));
+        }
         return "vehicle_records/vehicle-records-delete";
     }
 
     @PostMapping
-    public String processDelete(Long id, Long vehicleId){
-        VehicleRecord vehicleRecord = vehicleRecordService.getVehicleRecordById(id);
-        vehicleRecordService.removeVehicleRecord(vehicleRecord);
+    public String processDelete(Long vehicleRecordId, Long vehicleId){
+        Optional<VehicleRecord> optionalVehicleRecord = vehicleRecordService.getVehicleRecordById((vehicleRecordId));
+        if(optionalVehicleRecord.isPresent()){
+            VehicleRecord vehicleRecord = optionalVehicleRecord.get();
+            vehicleRecordService.removeVehicleRecord(vehicleRecord);
+        }
         return "redirect:/vehicle-records/list?id="+vehicleId;
     }
 }
